+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - Added a configurable timeout per case.
 - Changed default stats interval to 120s
+- Added seed loading/persistence to a directory
 
 * 1.0.0-beta.2